Adding per-stream additional columns to display GELF fields.
Custom GELF fields are one of the more powerful parts of Graylogs,
and they should be highlighted.  This change allows the user to
optionally specify additions columns for the messages table display
so the GELF fields are visible.

In my use-case, I have a 'production warnings' stream that collects
warnings from a number of different production services, and I want
to be able to view which service is emitting the errors at a glance.