Improving stream alarm email messages.
Two quick improvements to the streamalarm messages:
 * Adding a link back to the Graylog2 instance
 * Displaying a configurable number of recent messages

This is addresses http://jira.graylog2.org/browse/WEBINTERFACE-67.
This should probably be refactored further, into a mailer, etc, but
these two changes add a lot of context to the email message and
represent a signifcant improvement.

One thing to consider is that alerts are often viewed on smartphones,
so it is important to include context and add easy links.