de-staticize many many many classes.
Introduced a GraylogServer class which takes over most of the responsibility of the old Main.
GraylogServer contains everything pertaining to the application, without using static access.
Thus I needed to "inject" the server instance into basically all threads to get rid of the static method calls and singletons.

There are many more opportunities to avoid creating objects for each message, this hasn't been done yet.
also the encapsulation of the mongo db connection is dicy. big TODO there.

There's more work to be done to not exit if some external services are not up yet.