Don't create a PID file if asked not to do so
Systems administrators and, more importantly, packagers, may wish to handle PID
files by themselves, using tools such as start-stop-daemon or equivalents. In
this scenario, having graylog2 create a PID file is undesirable.

Add an option so that graylog2 not attempt to create a PID file. This allows
more graceful cooperation with such tools.