GELFMessage: more liberal ZLIB acceptance
If a compression level other than default is used, or if a dictionary
is not, GELFMessage.Type.determineType would reject the incoming
message as unsupported, even though Java's InflaterInputStream can
handle the data just fine.

This changes determineType to use the simple algorithm described in
the RFC to determine if a message is zlib compressed or not, once the
first byte is matched.