Measure buffer sizes
A set of AtomicInteger bound to the Core object that are incremented and decremented on buffer read and write allow us to indicate current utilization of the buffers. Using Metrics was not an option here, because the most precise calculation (one minute average) was way to deferred. The new CLI parameter -s/--statistics enables periodical printing of buffer utilization to STDOUT. Next step is to write this information to MongoDB and display it in the web interface.

fixes #SERVER-199