optimized stream rule type conversion
the old behaviour was to always try to run parseInt without taking a closer look if this is necerssary at all. even worse: it was kind of expecting exceptions if a user tried to match greater/smaller on a non-numerical field in a message. this would have lead to a lot (!) of expensive exceptions. @kroepke and @joschi might want to take a look because they have more experience with Java

next step is to order execution of streamrules better: the more expensive regex matcher should always be the last in the list. the rules are AND connected and there is no sense in running them first if fast comparisons will fail later.