Minor tweaks for stream rules editor
Including ugly hack that includes patching of the minified bundled
bootstrap javascript because the typeahead plugin is doing a
stopPropagate for the keyup/change events that prevents them from
bubbling up the DOM, so the natural language preview in the stream rules
add modal is not working. I have removed this from the bundled
bootstrap.js for now, which is not pretty. An alternative would be to
create a jQuery plugin which is attached during the creation of each
form.