Testing stream rules could not be applied to all message fields
The API was returning a filtered set of fields for messages retrieved.
Adding method to return filtered fields for other methods that rely on
it.
Keeping full set of fields in MessageResult and filtering fielfs in
accessor methods insteaf of constructor.
Adding new route to API that returns filtered set of fields, returning
full set of fields per default.

Fixes #528