Giving some love to the testing situation
Adding TestNG and converting all tests to it (besides one, which is
        currently ignored)
Adding some test classes for the AlertCondition classes
Adding Mockito and PowerMock to test dependencies

Currently both JUnit and TestNG suites are supported and run. This is
for legacy. If we are happy with TestNG we can disable JUnit completely.