Disabling check for open files limit if notification is already present
This prevents log spamming on the server and flickering of notifications
if the open file limit is too low on multiple hosts. This is not the
optional solution for it, because we need the ability in the Notification
class to check for a tuple of notification type and node, right now it
checks for the type only.