Including configurable backlog of messages in stream alert notification
This change adds a parameter backlog (defaulting to 0) to stream alert
conditions which determines how many relevant messages will be included
in the stream alert noficationt that is sent out to configured
subscribers. A class MessageFormatter which renders the supplied
messages to plain text is used, which can be modified in the future to
support a user-supplied template.

Fixes #431
(At least basically. Some features like the user templates and the links
 to the web interface will be implemented in the near future.)