Moving output message batching from OutputBufferProcessor to ES output
This fixes an issue where on low traffic situations messages are being
left in the buffer of OutputBufferProcessors because commits of the
buffer happen only when reaching the batch size (for reasonable batch
sizes this condition is practically never met) or when the
disruptor is signalling an end of batch. Messages of a disruptor work
batch do not have to end up in the same OutputBufferProcessor, so
messages which are the beginning of a disruptor work batch are left in
the buffer until the next messages comes in at this processor with the
end of batch flag set.

Fixes #554