move graylog2-rest-routes into server repository
having it external means that it needs to access maven to retrieve the latest server code, which can easily go wrong if the latest snapshot isn't availabe (e.g. on Travis)
this way the jar is versioned the same way as the server, and will always stay up to date. It conceptually belongs to the server's REST API anyway.

One can use `mvn install` on travis ci to gain access to the correct snapshot jar when building clients like the web interface.