Fixing type information for created alarm callbacks.
When a user submits a form for a new alarm callback, the data is begin
transmitted form-encoded, so basically without any type information.
To solve this, we are taking a look at the alarm callback type we are
about to add and cast the fields to proper types according to the
alarm callback specification, before POSTing the data to the server.