Adding ObjectId serializer to ObjectMapper, adding exception handlers
This allows us to return ObjectIds which are properly serialized by
Jackson.

Splitting up server/radio binding for ObjectMapper, because the ObjectId
serializer is not available (and not needed) in the radio module.
Adding exception handlers to Jersey for our own NotFoundException and
ValidationException, so we can just throw instead of wrap them.