Adding basic support for outputs.
Outputs are created globally and assigned to streams individually.
This means a single output can be associated with multiple streams, just like inputs.

To reproduce these associations, the logic on the server is split in two parts.
The OutputResource is used to manage outputs themselves, the StreamOutputResource
is used in order to manage associations between streams and outputs.