Removing hostname lookup during message initialization for now.
The getByName call has an unpredictable runtime and is therefore
not applicable in the critical path of message processing. A non-
functioning or overloaded resolver could otherwise slow down message
processing massively.

Nonetheless the reverse ptr of the source ip is valuable information
that we want to make available internally or for scripting. Therefore
we need to think of way to retrieve it during later processing steps
or lazily.