Basic support for rest controllers coming from plugins.
This is done using a static REST resource that looks up the resource
by Path. This is not a good solution because it basically does what
Jersey (or any other JAX-RS implementation) should do.
Unfortunately adding the REST resources coming from plugins as child
resources to the ResourceConfig object did not work properly.