Fix handling of parameterized types in Swagger Generator
Instead of failing with a ClassCastException, the Generator class now simply
drops the generics information on a ParameterizedType and just handles the
raw type of the class.

There might be a better and more correct way of doing this (e. g. with Guava
Reflection, see https://code.google.com/p/guava-libraries/wiki/ReflectionExplained)
but this works for now.