Consolidate sbt build
We got rid of the 'package-snapshot' command and can now use 'universal:package-zip-tarball'
directly. The name of the artifact is overridden in case of a SNAPSHOT build and the example
config will be mapped into the archive.

Selfie of the author of this changeset during basically every stage of development and
contact with SBT:

http://31.media.tumblr.com/bcaca0957cae39e4fd6fa056167f3603/tumblr_n8eup6Xzjv1s373hwo1_400.gif