Freeing up UDP handlers as fast as possible.
This is supposed to help decreasing the number of dropped UDP packets
during load peaks. Incoming packets are being added to a blocking queue
on which workers are operating. Making this step async helps freeing
threads handling incoming packets, so less packets are dropped.

This will become obsolete when message spooling is implemented.