Unbreak outputs API endpoint.
* Return `id` instead of "_id".
* Ensure correct serialization of the "id" field.

The rest client now expects an "id" field instead of "_id". This is
probably due to the port from Gson to Jackson for JSON.

The serialization of the ObjectId broke with commit 3f7ef53 when the
JerseyObjectMapperProvider got replaced by ObjectMapperProvider.
ObjectMapperProvider does not serialize the ObjectId correctly because
it is also used in graylog2-radio where no MongoDB library is available.

Instead of reverting the JerseyObjectMapperProvider change I converted
the ObjectId in OutputImpl.asMap() to a hex string because other data
models do the same.
It is not really nice this way but at least consistent with the other
models.

(cherry picked from commit f5b7d46)