Fix handling of IPv6 addresses (and hosts) in mongodb_replica_set
The previous implementation of Configuration#getMongoReplicaSet relied on a simple
host/port format which would break with IPv6 addresses. The new implementation is
using HostAndPort from Google Guava to parse the host/port literals (now with optional
default port 27017) and convert them to ServerAddress for the Mongo Java Driver.

Fixes #751