Ensure correct temporal order in NaturalDateParser.Result
Some ambiguous natural date strings like "last one hour" could lead to a wrong
order of 'from' and 'to' in NaturalDateParser.Result. This changeset sorts the
resulting list so that the lowest date is always first and thus being used as
'from' attribute.

Ultimately this should be fixed (or at least considered to be fixed) in natty
itself.

Fixes Graylog2/graylog2-web-interface#982