Add ContentPackLoaderPeriodical to load bundled content packs
The ContentPackLoaderPeriodical loads content packs shipped with Graylog2 into
the database if they don't exist. This serves as mechanism to initially seed the
database on the first startup of the server. Existing content packs will not be
overwritten.

The content packs must be placed in ./graylog2-server/src/main/resources/contentpacks/
as their JSON representation (same format being used to import content packs) and are
automatically recognized if they have the extension ".json".

Refs Graylog2/graylog2-web-interface#935