Adding ability to add config beans in plugins.
This allows plugins to be configurable from the main Graylog2 config
file using plugin-defined configuration options.

To do this, the bootstrapping is now slightly different. Initially, the
configuration file is read for the plugin directory setting only, before
all plugins are loaded and their config beans are extracted. Those are
then added to the main JadConfig instance just before the configuration
file is processed completely.

This has been added to the Bootstrap commands, so it does not work for
the old separated artifacts.