replace kafka's internal scheduled log maintenance jobs by java versions
this is to be able to use our customize algorithm to determine when we can let go of a journal segment.
essentially we don't want to use too much disk space if we don't need to, in normal operation the journal should be only one segment, not the maximum amount of segments.
the default kafka behavior makes sense for a message queue system, but not so much for a temporary buffer.

brought to you by project shavedyak