fix race between setting index read only and indexing into deflector alias
we are waiting 30 seconds to set the index to read only and trigger optimization afterwards to allow other nodes in the cluster to finish their indexing operations which might still target the old deflector index.

we really need to coordinate indexing and deflector management clusterwide, this is just trying to mask the symptoms.

fixes #833