Stubbing services, adding security context. deduplicating REST resources
More work to share REST controller classes between radio and server.
Therefore we are supplying a special security context that is providing
a special "allow everything" subject and we are stubbing user services.
Moving DTOs used to rest-model module, so it can be used as an API
return type later.
Extracting several semantically associated methods of the SystemResource
to separate resources.