Do not store the structured data in the "message" field.
If there is structured stat in the syslog message, the structured data
and the message ID should not be in the "message" field.

There is still a problem if the syslog STRUCTURED-DATA field is a
NILVALUE ("-"). Then the message plus the "-" for the nil value will end
up in the "message" field. This needs to be solved in the syslog4j
library.

Refs: Graylog2/graylog2-server#845