Change order of event bus register/unregister in ThrottleableTransport.
Calling unregister twice causes an IllegalArgumentException. Since it might
be called in doStop() as well, make sure it is called after doStop() in
ThrottleableTransport and protect it with a try/catch block.

Execute doLaunch() before registering to the event bus because there
might be code that needs to be run before registering to the event bus.

Also check if throttling is enabled to updateThrottleState() because
the object might be registered on the event bus from the subclass.

Fixes #874.