Stream throughput now returns all nodes' combined throughput.
Before this was only returning the stream throughput on the node the web
interface is talking to. Unfortunately this introduces a lot more
requests being done periodically when running on a big cluster. It also
means that the throughput fetching is much more fragile because in our
code a single non-functioning node (happens especially during node
startup/shutdown) makes the whole request fail.

Fixes #1047