Create MessagePack singleton provider to avoid PermGen leak.
Prepare the MessagePack instance with a RadioMessage because generating
code for the MessagePack template registry is not thread-safe.

Before this, a new MessagePack instance got created for each processed
message which eventually lead to PermGen exhaustion. Not sure why the
generated classes haven't been garbage collected. (bug in msgpack?)

Fixes #907.