Use English locale to format numbers on helper
As we don't provide any kind of localisation at the moment, it's better
to not rely on user's locale to format the numbers. I also removed the
conversion of that returned String into a Double, as we shouldn't rely
on the String to be converted without a problem into a Double.

Fixes #1062