Quite some effort for a (still ugly) solution to mask passwords.
In the long run we need to store all entities configurations in typed
objects which are also used to generate the configuration request, so
everything is in one place. For now the filtering logic is extracted to
a separate helper class and return types for both resources are unified.

Fixes #1066