Mark optional MongoDB stats as @Nullable
Some stats from MongoDB are optional or are simply not implemented in older
versions. Unfortunately the docs are quite silent about that so we had to
learn this the hard way.

By comparing the outputs of MongoDB 2.2.7 on Ubuntu 14.04 the stats in question
have now been marked as optional by adding a @Nullable annotation.

Links to the MongoDB commands in question:

* http://docs.mongodb.org/manual/reference/command/buildInfo/
* http://docs.mongodb.org/manual/reference/command/dbStats/
* http://docs.mongodb.org/manual/reference/command/hostInfo/
* http://docs.mongodb.org/manual/reference/command/serverStatus/

Fixes #941