Respecting rest_transport_uri when generating entity URLs.
Adding helper method in REST resource base class which constructs a
UriBuilder pre-populated with the URI defined in rest_transport_uri.
If rest_transport_uri is not defined, it is either taken from
rest_listen_uri or picked from the first external network interface.

This helps us to generate proper Location headers for newly created
entities and allows the user to override the base URI if the Graylog
server/radio is run behind a proxy that is changing the scheme, host
name or URI of the request.

Fixes #1020