Add simple cluster event bus based on MongoDB and EventBus
The cluster event bus can be used to send events from one Graylog node to
all other Graylog nodes (e. g. to trigger a configuration reload).

Underneath the cluster event bus is using a capped collection in MongoDB
which is regularly polled by each Graylog node. Once the document has been
processed by a particular node, its node ID is being added to the "consumers"
array of the document so that it won't read it again. The deserialized payload
of a cluster event is being posted to the local server EventBus so that other
parts of the node can react upon it.

Every event being sent to the cluster is being wrapped inside a ClusterEvent
and *must* be serializable and deserializable by Jackson, otherwise the
receivers won't be able to deserialize the payload of the cluster event!