Change ClusterEventService into Periodical
Re-implementing ClusterEventService as Periodical is necessary for getting rid of
capped collections. Instead of a tailable cursor, we're now polling the "cluster_events"
collection every second for new cluster events.

While this is less efficient than using a capped collection and a tailable cursor,
we can keep track of consumers having processed the cluster events by updating the
"consumers" array in those documents.