Fixing bug where opening the edit dialog needed a second click.
Instead of generating the population form on demand (when clicking)
we are now rendering it straight away. It was done on demand when we
needed to do a request to fetch the configuration details. Now that
these are passed in for all types, we can just render the form.

(cherry picked from commit 3192679b633c8fd8275fe05ea99f36596e3b5a96)