Increase default RecvBufferSize for UDP transports to 64K
While the receive buffer size for UDP-based transports could be configured for each transport,
the effective size for the receive buffer was fixed to 8192 bytes by using a FixedReceiveBufferSizePredictor
with a hard-coded value.

By changing the default receive buffer size for UDP-based transports to 65535 bytes and using the
configured receive buffer size as value for the FixedReceiveBufferSizePredictor, the maximum UDP packet
size that can be processed by UDP-based transports is now finally configurable.

As the practical maximum for a UDP packet is 65507 bytes, the default of 65535 bytes for the receive buffer
should be fine (at the cost of increased memory allocations).

Fixes #1143