Fixing incorrect alert backlog population and size calculation.
This affects only field value conditions. Range is now passed in when
construction search queries, so null check for range is false and filter
is appended.
This is the least intrusive fix for this issue, we should take a look
at the relevant code parts for 1.1.0 and make the filter generation code
code more robust.

Conflicts:
	graylog2-server/src/main/java/org/graylog2/alerts/types/FieldValueAlertCondition.java