Change write concern for cluster related entities to FSYNCED
The semantics of WriteConcern.MAJORITY have been changed in MongoDB 2.6. Prior
to that version, this write concern produces an error in non-replica set setups.

Since the cluster config and cluster event collections aren't heavily used in
Graylog, it's safe to use the WriteConcern.FSYNCED (i. e. highest guarantee for
writing to master node) for mutating operations on those.

Fixes #1138