Unbreak index range calculation.
When getting the first message timestamp of an index for the "start"
parameter of the index range with "Searches.firstOfIndex()", a DESC
sort order was used. So this was actually getting the last message
timestamp of an index, not the first.

Change "Searches.firstOfIndex()" to use ASC sort order to fix this.

Also remove unused "Searches.lastOfIndex()" method.