Fixing alarm callbacks: correcting collection name, supporting isodates.
This is fixing a nasty bug introduced when replacing the plain mongo
driver alarm callback service and persisted model class with ones based
on AutoValue & MongoJack. During this step, the wrong collection name
was used. Also the format of the persisted mongo document has changed
for the created_at field. Before it was an ISODate() field, now it is a
plain string. Due to this, reading back of the old data failed. This
commit changes the type of the field read back and converts it to a
DateTime in the JSON creator method. With this, reading back both old
and new entities should be possible.

Fixes #1214