Encapsulate npm installs on different branches
Using npm on different branches is really annoying, as we have different
dependencies and we need to delete the `node_modules` folder and do `npm
install` every time that we switch branches.

Using a shell script to handle multiple `npm installs` and combining it
with the existing `gulp watch` task, I'm hoping we can make this easier.
Please use `npm run watch` to give it a try.