Executing integration tests sequentially for now.
Right now we are keeping and modifying test state in the MongoDB and
elasticsearch sources used by the graylog server the tests are run
against. These are modified during test setup, so any concurrent tests
would tamper with the state of other tests. Therefore we are running
integration tests sequentially now.