Handle exceptions in the JournallingMessageHandler.
Before this, an exception from the journal would leave invalid
RawMessageEvent objects in the batch List because the Converter sets
all fields to null after transforming the objects to Journal.Entry
objects. The batch List was not cleared in this case.
With the next endOfBatch the Coverter tried to transform the broken
objects again and threw a NPE because the fields are null.

This caused exceptions for every new event coming in and eventually
filled up the memory because the unbounded batch List was never cleared.

(cherry picked from commits f8d9391 6623e80 c1e45df)

Refs #1286