Allow startup without working Elasticsearch cluster
The previous implementation of IndexerSetupService (and related classes) would should down Graylog if
Elasticsearch wasn't available or the cluster health state was RED on startup. This made sense as long
as Graylog didn't have a proper on-disk journal in which messages could be persistently buffered.

The new implementation simply acknowledges the unavailability of Elasticsearch and generates an
appropriate system notification (and log entries) but does not shutdown Graylog.

Fixes #1136