Add MessageListCodec interface.
It makes it possible to decode a list of Message objects from a
RawMessage instead of just a single one. This is useful for protocols
like NetFlow or collectd where one packet contains multiple messages.

The new interface is needed to keep the Codec interface compatible. A
breaking change is only possible with the next major version.

Adjust DecodingProcessor and ProcessBufferProcessor to handle a list of
Message objects.