Initialise persisted widgets with wrong timeranges
When a persisted widget contains an invalid timerange, initialise the
widget nonetheless, so it can be edited or deleted from the web
interface.

In that case, an error will be logged into the graylog server logs, and
an exception will be raised when computing the value for the affected
widgets.

Fixes Graylog2/graylog2-web-interface#1505