Relax permissions to manage stream alerts
On Graylog 1.1 and 1.2, the manage stream alerts button was only displayed
for users that had edit permissions on a stream, and `output:read`
permissions. We are relaxing the permissions needed, to go back to the
ones we used in 1.0. Users with edit permissions for that stream will be
able to manage stream alerts.

Fixes #1659

(cherry picked from commit dbe7b28774fd1f59b3cd9cef4c78c00c4fceb2e6)