Initialize widgets during all cases of dashboard instantiation.
With the removal of the dashboard registry, dashboards loaded from the
persistence engine need to initialize the widgets from their embedded
fields every time they are instantiated. With the registry in place this
was done once. Now it needs to be done for both the all() and the load()
call.