Initial version of ClusterMetricsResource.
Due to the web interface now talking to a single node only (in contrast
to the play web interface component connecting to every node in the
cluster) we need a mechanism to retrieve metrics (and other resources)
from other nodes than the one we are talking to. Therefore a
ClusterMetricsResource was created that proxies requests to other nodes
for metrics retrieval.