Create shared and vendor bundles, add module ids, include in template.
This allows plugins to consume the vendor and shared bundles, keep
module ids static and reference them in the generated index.html
boilerplate.

Also adds `webpack.combined.config.js` which is used for `npm start` to
include plugins. To enable any plugin with a web part in a development
run, just symlink the plugin directory in the `plugins` directory.