Removing shared module, passing Store/ActionsProvider in context.
This allows us to share singleton instances of stores and actions
throughout all components of the web interface and all plugins.

The shared bundle has been removed, because it prevents hot reloading of
shared components. Components which are reused by plugins are bundled
into plugins instead of referenced from the shared bundle. This has the
downside that plugins need rebundling when components of the web
interface are modified (because of bugfixing or enhancements) but it
also makes them resilient to API changes.