Return 200 for OPTIONS requests for non-existent resources.
Before, an OPTIONS requests for a non-existent resource would return a
404, just like the request itself. This is confusing for XMLHTTP
requests from browsers, because its indistinguishable on the upper
layers if the pre-flight request or the request itself failed. Therefore
we are returning a 200 for the request now, as the pre-flight request
itself is actually ok, even if the actual request is not going to be.