Show modal when server connection breaks/is not working.
Before this change, a non-working server connection was indicated only
by javascript errors in the browser console and neverending spinners.
With this change, whenever requests are unsuccessfull in a fundamental
way, or the occasional polling for the server does not succeed, an
overlay will be rendered indicating the error.

Right now the design is very basic and questionable, additional styling
and displaying of the error can be added later.