working pipeline intepreter
The interpreter depends on a few data structures:
 - rules collection and fully loaded rule set in memory (cached + updated via event bus)
 - pipeline collection and loaded set in memory (cached + updated via event bus)
 - pipeline -> stream assignment (cached + updated via event bus)

Pipelines are not being run until they are attached to a stream. It uses a "stream id" of 'default' to identify incoming messages which do not have a stream assigned yet.

The pipeline updater resolves all functions, rules and stages inside pipelines, this is done by name and not internal ID.

Execution is as follows:
For each incoming message the interpreter determines which pipelines can possibly run, based on the streams a message is assigned to.
For each of the pipelines their stages will be executed, in ascending stage number order. Stages with the same order number are being run after each other.
For each of the stages their rules' conditions are evaluated and gate which rules' actions are being run.
The match clause determines whether the subsequent stages for that particular pipeline will be run.
Each message is being run through a pipeline once per stream.
After all pipelines have been run, the interpreter determines whether additional pipelines need to be run as well, based on the new stream assignments.
The loop stops when no more new streams have been assigned to a message.

Note: This implementation does not yet guarantee that stages are isolated from each other while they run.