Add Maven build profile for Travis CI
Travis CI ensures only 2 cores and 3 GB of memory per build job. Unfortunately it's also using
Docker so that the number of "visible" CPU cores (e. g. via /proc/cpuinfo) is higher than the
actual number of available cores, so that the maven-surefire-plugin will try to use all of those
cores for running tests.

Explicitly setting the number of parallel forks for running tests to 2 in the Maven build profile
"travis" should work around this issue.

https://docs.travis-ci.com/user/common-build-problems/#My-build-script-is-killed-without-any-error