Allow overriding web if server uri from X-Graylog-Server-URL header.
This allows http clients (including proxies) to define the server uri
that is used by the web interface code to contact the REST API.

This is especially useful for reverse proxies exposing internal Graylog
servers to the outside. In the proxy configuration you can just set the
X-Graylog-Server-URL that is passed to the backend to a URL that is
visible from outside enabling access to the server from the web if.