Overhaul function/param descriptors and precompute constant function arguments
To avoid reevaluating function arguments which never change for each message and are potentially expensive to calculate (either because their expression tree is deep or the function needs to construct other objects)  the Function interface now contains a callback to provide precomputed values for each constant argument.

This callback is only invoked for provably constant expressions, so it is safe to implement for any function.
The recommended way is to inherit from AbstractFunction<T> which implements the straightforward precompute function by simply memoizing the evaluated value.
For special cases, like the RegexMatch function, which precomputes a Pattern.compile from a constant parameter descriptors take a java.util.function.Function now, which gets invoked with the computed value. The default is the identity function.

Parameter descriptors are now responsible for evaluating argument expressions, and are fully generic.