Log response body message in console on error
We shouldn't show those messages to the user in an error notification,
but make them available somehow. Ideally there would be a 'show
additional information' in the error dialog, but we can't do that. For
now we log them into the console and remove code showing them to the
user in notifications.